Initial fold-in of expanded prompt files.
