+++ go/plugins/dotprompt/genkit.go
@@ -12,7 +12,6 @@ import (
 	"strings"
 
 	"github.com/firebase/genkit/go/ai"
-	"github.com/firebase/genkit/go/ai/prompt"
 	"github.com/firebase/genkit/go/core/tracing"
 	"github.com/firebase/genkit/go/genkit"
 )
@@ -182,9 +181,9 @@ func (p *Prompt) Register(g *genkit.Genkit) error {
 		g,
 		"dotprompt",
 		name,
-		prompt.WithMetadata(metadata),
-		prompt.WithInputType(p.Config.DefaultInput),
-		prompt.WithRender(p.buildRequest),
+		ai.WithMetadata(metadata),
+		ai.WithInputType(p.Config.DefaultInput),
+		ai.WithRenderFn(p.buildRequest),
 	)
 	return err
 }
@@ -193,7 +192,7 @@ func (p *Prompt) Register(g *genkit.Genkit) error {
 // passes the rendered template to the AI model specified by
 // the prompt.
 //
-// This implements the [ai.Prompt] interface.
+// This implements the [ai.PromptDef] interface.
 func (p *Prompt) Generate(ctx context.Context, g *genkit.Genkit, opts ...GenerateOption) (*ai.ModelResponse, error) {
 	tracing.SetCustomMetadataAttr(ctx, "subtype", "prompt")
 	var pr PromptRequest
